refactor(processors): share message-condensing helpers across classifier, judge, and planner
